mgmtsystem_kpi migration to odoo 9.0
