+++ mgmtsystem_kpi/data/mgmtsystem_kpi.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data noupdate="1">
+        <record forcecreate="True" id="ir_cron_mgmtsystem_kpi_action" model="ir.cron">
+            <field name="name">Update KPI values</field>
+            <field name="user_id" ref="base.user_root"/>
+            <field name="interval_number">1</field>
+            <field name="interval_type">hours</field>
+            <field name="numbercall">-1</field>
+            <field eval="False" name="doall"/>
+            <field eval="'mgmtsystem.kpi'" name="model"/>
+            <field eval="'update_kpi_value'" name="function"/>
+            <field eval="'()'" name="args"/>
+        </record>
+    </data>
+</odoo>
