+++ mgmtsystem_kpi/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from . import mgmtsystem_kpi_category
+from . import mgmtsystem_kpi_threshold_range
+from . import mgmtsystem_kpi_threshold
+from . import mgmtsystem_kpi_history
+from . import mgmtsystem_kpi
