+++ mgmtsystem_kpi/models/mgmtsystem_kpi_history.py
@@ -1,0 +1,46 @@
+#  -*- encoding: utf-8 -*-
+##############################################################################
+#
+#    OpenERP, Open Source Management Solution
+#    Copyright (C) 2012 Savoir-faire Linux (<http://www.savoirfairelinux.com>).
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
+from openerp import fields, models
+
+
+class MgmtsystemKPIHistory(models.Model):
+    """History of the KPI."""
+
+    _name = "mgmtsystem.kpi.history"
+    _description = "History of the KPI"
+    _order = "date desc"
+
+    name = fields.Char('Name', size=150, required=True,
+                       default=fields.Datetime.now(),)
+    kpi_id = fields.Many2one('mgmtsystem.kpi', 'KPI', required=True)
+    date = fields.Datetime(
+        'Execution Date',
+        required=True,
+        readonly=True,
+        default=fields.Datetime.now()
+    )
+    value = fields.Float('Value', required=True, readonly=True)
+    color = fields.Text('Color', required=True,
+                        readonly=True, default='#FFFFFF')
+    company_id = fields.Many2one(
+        'res.company', 'Company',
+        default=lambda self: self.env.user.company_id.id)
