+++ mgmtsystem_kpi/views/mgmtsystem_kpi_category.xml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <!-- Configuration menu -->
+
+        <menuitem id="menu_mgmtsystem_configuration_kpi"
+            name="KPI"
+            parent="mgmtsystem.menu_mgmtsystem_configuration"
+            groups="mgmtsystem.group_mgmtsystem_manager"
+            sequence="20"/>
+
+	<!-- Categories -->
+
+        <record id="view_mgmtsystem_kpi_category_tree" model="ir.ui.view">
+            <field name="name">mgmtsystem.kpi.category.tree</field>
+            <field name="model">mgmtsystem.kpi.category</field>
+            <field name="arch" type="xml">
+                <tree string="Categories">
+                    <field name="name"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="view_mgmtsystem_kpi_category_form" model="ir.ui.view">
+            <field name="name">mgmtsystem.kpi.category.form</field>
+            <field name="model">mgmtsystem.kpi.category</field>
+            <field name="arch" type="xml">
+                <form string="Category">
+                    <field name="name"/>
+                    <newline/>
+                    <field name="description" colspan="4"/>
+                </form>
+            </field>
+        </record>
+
+        <record model="ir.actions.act_window" id="open_mgmtsystem_category_list">
+            <field name="name">Categories</field>
+            <field name="res_model">mgmtsystem.kpi.category</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree,form</field>
+            <field name="view_id" ref="view_mgmtsystem_kpi_category_tree"/>
+        </record>
+
+        <menuitem id="menu_mgmtsystem_configuration_kpi_category"
+            name="Categories"
+            action="open_mgmtsystem_category_list"
+            parent="menu_mgmtsystem_configuration_kpi"
+            groups="mgmtsystem.group_mgmtsystem_manager"
+            sequence="10"/>
+    </data>
+</odoo>
