+++ mgmtsystem_kpi/views/mgmtsystem_kpi_threshold.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+	<!-- Thresholds -->
+        <record id="view_mgmtsystem_kpi_threshold_tree" model="ir.ui.view">
+            <field name="name">mgmtsystem.kpi.threshold.tree</field>
+            <field name="model">mgmtsystem.kpi.threshold</field>
+            <field name="arch" type="xml">
+                <tree string="Thresholds">
+                    <field name="name"/>
+                    <field name="invalid_message"/>
+	            <field name="company_id" group="base.group_multi_company"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="view_mgmtsystem_kpi_threshold_form" model="ir.ui.view">
+            <field name="name">mgmtsystem.kpi.threshold.form</field>
+            <field name="model">mgmtsystem.kpi.threshold</field>
+            <field name="arch" type="xml">
+                <form string="Threshold">
+                    <field name="name"/>
+                    <newline/>
+                    <separator string="Ranges" colspan="2"/>
+                    <newline/>
+                    <field name="range_ids" nolabel="1" colspan="2"/>
+                    <newline/>
+                    <separator string="KPIs" colspan="2"/>
+                    <newline/>
+                    <field name="kpi_ids" nolabel="1" colspan="2"/>
+                    <newline/>
+                    <field name="invalid_message" nolabel="1" attrs="{'invisible' : [('invalid_message', '=', '')]}" colspan="2"/>
+                    <newline/>
+	            <field name="company_id" group="base.group_multi_company"/>
+                </form>
+            </field>
+        </record>
+
+        <record model="ir.actions.act_window" id="open_mgmtsystem_threshold_list">
+            <field name="name">Thresholds</field>
+            <field name="res_model">mgmtsystem.kpi.threshold</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree,form</field>
+            <field name="view_id" ref="view_mgmtsystem_kpi_threshold_tree"/>
+        </record>
+
+        <menuitem id="menu_mgmtsystem_configuration_kpi_threshold"
+            name="Thresholds"
+            action="open_mgmtsystem_threshold_list"
+            parent="menu_mgmtsystem_configuration_kpi"
+            groups="mgmtsystem.group_mgmtsystem_manager"
+            sequence="10"/>
+	</data>
+</odoo>
