+++ mgmtsystem_kpi/views/mgmtsystem_kpi_threshold_range.xml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+	<!-- Ranges -->
+        <record id="view_mgmtsystem_kpi_threshold_range_tree" model="ir.ui.view">
+            <field name="name">mgmtsystem.kpi.threshold.range.tree</field>
+            <field name="model">mgmtsystem.kpi.threshold.range</field>
+            <field name="arch" type="xml">
+                <tree string="Ranges">
+                    <field name="name"/>
+                    <field name="min_value"/>
+                    <field name="max_value"/>
+                    <field name="color" widget="color"/>
+                    <field name="invalid_message"/>
+	            <field name="company_id" group="base.group_multi_company"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="view_mgmtsystem_kpi_threshold_range_form" model="ir.ui.view">
+            <field name="name">mgmtsystem.kpi.threshold.range.form</field>
+            <field name="model">mgmtsystem.kpi.threshold.range</field>
+            <field name="arch" type="xml">
+                <form string="Range">
+                    <field name="name"/>
+                    <field name="color"/>
+	            <field name="company_id" group="base.group_multi_company"/>
+                    <newline/>
+                    <separator string="Minimum" colspan="4"/>
+                    <newline/>
+                    <field name="min_type"/>
+                    <field name="min_fixed_value" attrs="{'invisible' : [('min_type', '!=', 'static')]}"/>
+                    <field name="min_dbsource_id" attrs="{'invisible' : [('min_type', '!=', 'external')]}"/>
+                    <newline/>
+                    <field name="min_code" colspan="4" attrs="{'invisible' : [('min_type', 'NOT IN', ('local','external','python'))]}"/>
+                    <newline/>
+                    <separator string="Maximum" colspan="4"/>
+                    <newline/>
+                    <field name="max_type"/>
+                    <field name="max_fixed_value" attrs="{'invisible' : [('max_type', '!=', 'static')]}"/>
+                    <field name="max_dbsource_id" attrs="{'invisible' : [('max_type', '!=', 'external')]}"/>
+                    <newline/>
+                    <field name="max_code" colspan="4" attrs="{'invisible' : [('max_type', 'NOT IN', ('local','external','python'))]}"/>
+                    <newline/>
+                    <separator string="Thresholds" colspan="4"/>
+                    <field name="threshold_ids" nolabel="1" colspan="4"/>
+                    <field name="invalid_message" nolabel="1" attrs="{'invisible' : [('invalid_message', '=', '')]}" colspan="4"/>
+                </form>
+            </field>
+        </record>
+
+        <record model="ir.actions.act_window" id="open_mgmtsystem_threshold_range_list">
+            <field name="name">Ranges</field>
+            <field name="res_model">mgmtsystem.kpi.threshold.range</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree,form</field>
+            <field name="view_id" ref="view_mgmtsystem_kpi_threshold_range_tree"/>
+        </record>
+
+        <menuitem id="menu_mgmtsystem_configuration_kpi_range"
+            name="Ranges"
+            action="open_mgmtsystem_threshold_range_list"
+            parent="menu_mgmtsystem_configuration_kpi"
+            groups="mgmtsystem.group_mgmtsystem_manager"
+            sequence="20"/>
+
+        <menuitem id="menu_mgmtsystem_configuration_kpi_dbsource"
+            name="Data Sources"
+            action="base_external_dbsource.action_dbsource"
+            parent="menu_mgmtsystem_configuration_kpi"
+            groups="mgmtsystem.group_mgmtsystem_manager"
+            sequence="20"/>
+    </data>
+</odoo>
